Improve special item display
