+++ SpecialItems/src/main/java/com/specialitems/util/Tagger.java
@@ -1,7 +1,6 @@
 package com.specialitems.util;
 
 import com.specialitems.leveling.Keys;
-import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -28,9 +27,6 @@ public final class Tagger {
 
         meta.setUnbreakable(true);
         try { meta.addItemFlags(ItemFlag.HIDE_UNBREAKABLE, ItemFlag.HIDE_ENCHANTS); } catch (Throwable ignored) {}
-        if (meta.getEnchants().isEmpty()) {
-            try { meta.addEnchant(Enchantment.UNBREAKING, 1, true); } catch (Throwable ignored) {}
-        }
 
         item.setItemMeta(meta);
     }
